// src/types.ts
export interface PaymentRequest {
  /**
   * Amount that will be debited from the payer account
   */
  amount: string;

  /**
   * ISO4217 Currency
   */
  currency: string;

  /**
   * External id is used as a reference to the transaction.
   * External id is used for reconciliation.
   * The external id will be included in transaction history report.
   * External id is not required to be unique.
   */
  externalId: string;

  /**
   * Party identifies a account holder in the wallet platform.
   * Party consists of two parameters, type and partyId.
   * Each type have its own validation of the partyId
   *   - MSISDN - Mobile Number validated according to ITU-T E.164. Validated with IsMSISDN
   *   - EMAIL - Validated to be a valid e-mail format. Validated with IsEmail
   *   - PARTY_CODE - UUID of the party. Validated with IsUuid
   */
  payer: Payer;

  /**
   * Message that will be written in the payer transaction history message field.
   */
  payerMessage: string;

  /**
   * Message that will be written in the payee transaction history note field.
   */
  payeeNote: string;

  /**
   * URL to the server where the callback should be sent.
   */
  callbackUrl?: string;
}
// ...
export interface Payer {
  partyIdType: PartyIdType;
  partyId: string;
}

export type PartyIdType = "MSISDN" | "EMAIL" | "PARTY_CODE";